backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
redis==5.0.8
orjson==3.10.7
pydantic==2.9.2
jinja2>=3.1.0

